Rename `Worker` -> `Initializer`
